+++ src/main/java/com/project/demo/logic/entity/crop/CropRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CropRepository extends JpaRepository<Crop, Integer> {
+public interface CropRepository extends JpaRepository<Crop, Long> {
 }
