+++ src/main/java/com/project/demo/logic/entity/equipment/EquipmentMachineryRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface EquipmentMachineryRepository extends JpaRepository<EquipmentMachinery, Integer> {
+public interface EquipmentMachineryRepository extends JpaRepository<EquipmentMachinery, Long> {
 }
