+++ src/main/java/com/project/demo/logic/entity/farmequipment/FarmEquipmentRepository.java
@@ -7,7 +7,7 @@ import java.util.List;
 
 @Repository
 public interface FarmEquipmentRepository extends JpaRepository<FarmEquipment, FarmEquipmentId> {
-    List<FarmEquipment> findById_FarmId(Integer farmId);
+    List<FarmEquipment> findById_FarmId(Long farmId);
 
-    List<FarmEquipment> findById_EquipmentId(Integer equipmentId);
+    List<FarmEquipment> findById_EquipmentId(Long equipmentId);
 }
