+++ src/main/java/com/project/demo/logic/entity/managementaction/ManagementActionRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ManagementActionRepository extends JpaRepository<ManagementAction, Integer> {
+public interface ManagementActionRepository extends JpaRepository<ManagementAction, Long> {
 }
